requirements.txt
# Web framework
fastapi>=0.95.0
uvicorn[standard]>=0.23.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6

# Config & env
python-dotenv>=1.0.0

# Security
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.22.0
httpx>=0.24.0
pydantic[email]
pydantic-settings
python-multipart